mini_projects/day10_blog/requirements.txt
asgiref==3.10.0
attrs==25.4.0
Django==5.0.7
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Pillow==10.0.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
sqlparse==0.5.3
uritemplate==4.2.0
